renaming pull push buttons
